Addressed changes in review
- refactored common code for immutable types into its own base class
- move converter from *.hibernate to *.persistence package
- refactored dialect code from the two locations it is used in
- converted string settings names to symbolic constants
