+++ core/src/main/java/google/registry/persistence/ImmutableUserType.java
@@ -1,0 +1,45 @@
+// Copyright 2019 The Nomulus Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package google.registry.persistence;
+
+import org.hibernate.HibernateException;
+import org.hibernate.usertype.UserType;
+
+/** Base class containing common functionality for immutable user types. */
+abstract class ImmutableUserType implements UserType {
+  @Override
+  public boolean equals(Object x, Object y) throws HibernateException {
+    return x.equals(y);
+  }
+
+  @Override
+  public int hashCode(Object x) throws HibernateException {
+    return x.hashCode();
+  }
+
+  @Override
+  public Object deepCopy(Object value) throws HibernateException {
+    return value;
+  }
+
+  @Override
+  public boolean isMutable() {
+    return false;
+  }
+
+  @Override
+  public Object replace(Object original, Object target, Object owner) throws HibernateException {
+    return original;
+  }
+}
